+++ packages/api/lib/writeRecords/write-execution.js
@@ -2,6 +2,7 @@ const isNil = require('lodash/isNil');
 
 const {
   ExecutionPgModel,
+  translateApiExecutionToPostgresExecution,
 } = require('@cumulus/db');
 const {
   upsertExecution,
@@ -31,7 +32,7 @@ const Logger = require('@cumulus/logger');
 const { publishExecutionSnsMessage } = require('../../lib/publishSnsMessageUtils');
 const Execution = require('../../models/executions');
 
-const logger = new Logger({ sender: '@cumulus/sfEventSqsToDbRecords/write-execution' });
+const logger = new Logger({ sender: '@cumulus/api/lib/writeRecords/write-execution' });
 
 const shouldWriteExecutionToPostgres = ({
   messageCollectionNameVersion,
@@ -85,53 +86,50 @@ const buildExecutionRecord = ({
   });
 };
 
-const writeExecutionViaTransaction = async ({
-  cumulusMessage,
-  collectionCumulusId,
-  asyncOperationCumulusId,
-  parentExecutionCumulusId,
-  trx,
-  executionPgModel = new ExecutionPgModel(),
-  updatedAt,
-}) => {
-  const executionRecord = buildExecutionRecord({
-    cumulusMessage,
-    collectionCumulusId,
-    asyncOperationCumulusId,
-    parentExecutionCumulusId,
-    updatedAt,
-  });
-  logger.info(`About to write execution ${executionRecord.arn} to PostgreSQL`);
-  const upsertResponse = await executionPgModel.upsert(trx, executionRecord);
-  logger.info(`Successfully wrote execution ${executionRecord.arn} to PostgreSQL with cumulus_id ${upsertResponse[0]}`);
-  return upsertResponse;
-};
-
 const writeExecutionToDynamoAndES = async (params) => {
   const {
-    cumulusMessage,
+    dynamoRecord,
     executionModel,
-    updatedAt = Date.now(),
     esClient = await Search.es(),
   } = params;
-  const executionApiRecord = generateExecutionApiRecordFromMessage(cumulusMessage, updatedAt);
   try {
-    await executionModel.storeExecution(executionApiRecord);
+    await executionModel.storeExecutionRecord(dynamoRecord);
     await upsertExecution({
       esClient,
-      updates: executionApiRecord,
+      updates: dynamoRecord,
       index: process.env.ES_INDEX,
     });
   } catch (error) {
-    logger.info(`Writes to DynamoDB/Elasticsearch failed, rolling back all writes for execution ${executionApiRecord.arn}`);
+    logger.info(`Writes to DynamoDB/Elasticsearch failed, rolling back all writes for execution ${dynamoRecord.arn}`);
     // On error, delete the Dynamo record to ensure that all systems
     // stay in sync
-    await executionModel.delete({ arn: executionApiRecord.arn });
+    await executionModel.delete({ arn: dynamoRecord.arn });
     throw error;
   }
 };
 
-const writeExecution = async ({
+const _writeExecutionRecord = ({
+  dynamoRecord,
+  postgresRecord,
+  knex,
+  executionModel = new Execution(),
+  executionPgModel = new ExecutionPgModel(),
+  updatedAt = Date.now(),
+  esClient,
+}) => knex.transaction(async (trx) => {
+  logger.info(`About to write execution ${postgresRecord.arn} to PostgreSQL`);
+  const [executionCumulusId] = await executionPgModel.upsert(trx, postgresRecord);
+  logger.info(`Successfully wrote execution ${postgresRecord.arn} to PostgreSQL with cumulus_id ${executionCumulusId}`);
+  await writeExecutionToDynamoAndES({
+    dynamoRecord,
+    executionModel,
+    updatedAt,
+    esClient,
+  });
+  return executionCumulusId;
+});
+
+const writeExecutionRecordFromMessage = ({
   cumulusMessage,
   knex,
   collectionCumulusId,
@@ -141,33 +139,38 @@ const writeExecution = async ({
   updatedAt = Date.now(),
   esClient,
 }) => {
-  let id;
+  const postgresRecord = buildExecutionRecord({
+    cumulusMessage,
+    collectionCumulusId,
+    asyncOperationCumulusId,
+    parentExecutionCumulusId,
+    updatedAt,
+  });
   const executionApiRecord = generateExecutionApiRecordFromMessage(cumulusMessage, updatedAt);
-  await knex.transaction(async (trx) => {
-    const [executionCumulusId] = await writeExecutionViaTransaction({
-      cumulusMessage,
-      collectionCumulusId,
-      asyncOperationCumulusId,
-      parentExecutionCumulusId,
-      trx,
-      updatedAt,
-    });
-    await writeExecutionToDynamoAndES({
-      cumulusMessage,
-      updatedAt,
+  return _writeExecutionRecord(
+    {
+      dynamoRecord: executionApiRecord,
+      postgresRecord,
+      knex,
       executionModel,
       esClient,
-    });
-    await publishExecutionSnsMessage(executionApiRecord);
-    id = executionCumulusId;
-  });
-  return id;
+    }
+  );
+};
+
+const writeExecutionRecordFromApi = async ({
+  record: dynamoRecord,
+  knex,
+  executionModel = new Execution(),
+}) => {
+  const postgresRecord = await translateApiExecutionToPostgresExecution(dynamoRecord, knex);
+  return _writeExecutionRecord({ dynamoRecord, postgresRecord, knex, executionModel });
 };
 
 module.exports = {
   buildExecutionRecord,
   shouldWriteExecutionToPostgres,
   writeExecutionToDynamoAndES,
-  writeExecutionViaTransaction,
-  writeExecution,
+  writeExecutionRecordFromMessage,
+  writeExecutionRecordFromApi,
 };